Refresh 10

+++ include/level_misc_macros.h
@@ -2,7 +2,7 @@
 #define LEVEL_MISC_MACROS_H
 
 #define MACRO_OBJECT_WITH_BEH_PARAM(preset, yaw, posX, posY, posZ, behParam) \
-    (((yaw * 0x10 / 45) << 9) | (preset + 0x1F)), posX, posY, posZ, behParam
+    ((s16)((yaw * 0x10 / 45) << 9) | (preset + 0x1F)), posX, posY, posZ, behParam
 
 #define MACRO_OBJECT(preset, yaw, posX, posY, posZ) \
     MACRO_OBJECT_WITH_BEH_PARAM(preset, yaw, posX, posY, posZ, 0)
@@ -25,4 +25,4 @@
 #define TRAJECTORY_END() \
     -1
 
-#endif
+#endif // LEVEL_MISC_MACROS_H
